% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primitives.R
\name{adddays}
\alias{adddays}
\title{Add days to a date}
\usage{
adddays(x, n)
}
\arguments{
\item{x}{An atomic vector.}

\item{n}{Integer count.}
}
\value{
A Date vector.
}
\description{
Add days to a date
}
