% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primitives.R
\name{addweeks}
\alias{addweeks}
\title{Add weeks to a date}
\usage{
addweeks(x, n)
}
\arguments{
\item{x}{An atomic vector.}

\item{n}{Integer count.}
}
\value{
A Date vector.
}
\description{
Add weeks to a date
}
