% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primitives.R
\name{addyears}
\alias{addyears}
\title{Add years to a date}
\usage{
addyears(x, n, invalid = c("previous", "next", "missing", "error"))
}
\arguments{
\item{x}{An atomic vector.}

\item{n}{Integer count.}

\item{invalid}{How to handle an invalid resulting date.}
}
\value{
A Date vector.
}
\description{
Add years to a date
}
