% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primitives.R
\name{assertcols}
\alias{assertcols}
\title{Assert that columns exist (alias)}
\usage{
assertcols(data, cols)
}
\arguments{
\item{data}{A data.frame.}

\item{cols}{Character vector of column names.}
}
\value{
See \code{\link[=assert_cols]{assert_cols()}}.
}
\description{
Assert that columns exist (alias)
}
