% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primitives.R
\name{assertrange}
\alias{assertrange}
\title{Assert that a column falls within a numeric range}
\usage{
assertrange(data, column, lower, upper)
}
\arguments{
\item{data}{A data.frame.}

\item{column}{Name of a single column.}

\item{lower}{Lower bound.}

\item{upper}{Upper bound.}
}
\value{
\code{data}, invisibly, if the assertion passes.
}
\description{
Assert that a column falls within a numeric range
}
