% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primitives.R
\name{assertrows}
\alias{assertrows}
\title{Assert that a row-wise condition holds for every row}
\usage{
assertrows(data, condition)
}
\arguments{
\item{data}{A data.frame.}

\item{condition}{A logical expression evaluated in the context of \code{data}.}
}
\value{
\code{data}, invisibly, if the assertion passes.
}
\description{
Assert that a row-wise condition holds for every row
}
