% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primitives.R
\name{betweendates}
\alias{betweendates}
\title{Test whether a date falls within a range}
\usage{
betweendates(x, start, end)
}
\arguments{
\item{x}{An atomic vector.}

\item{start}{Start bound (inclusive).}

\item{end}{End bound (inclusive).}
}
\value{
A logical vector.
}
\description{
Test whether a date falls within a range
}
