% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verbs.R
\name{bind_rows}
\alias{bind_rows}
\title{Bind rows}
\usage{
bind_rows(..., id = NULL)
}
\arguments{
\item{...}{Data frames or a single list of data frames.}

\item{id}{Optional name for a source identifier column.}
}
\value{
A data.table.
}
\description{
Combine data frames by rows.
}
