% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primitives.R
\name{blanktona}
\alias{blanktona}
\title{Recode blank strings to \code{NA}}
\usage{
blanktona(x)
}
\arguments{
\item{x}{An atomic vector.}
}
\value{
\code{x} with blanks replaced by \code{NA}.
}
\description{
Recode blank strings to \code{NA}
}
