% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primitives.R
\name{ceilingdate}
\alias{ceilingdate}
\title{Round a date up to a unit}
\usage{
ceilingdate(x, unit = c("day", "week", "month", "year"))
}
\arguments{
\item{x}{An atomic vector.}

\item{unit}{Time unit to round to.}
}
\value{
A Date vector.
}
\description{
Round a date up to a unit
}
