% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primitives.R
\name{cleannames}
\alias{cleannames}
\title{Clean column names to a unique, syntactic form}
\usage{
cleannames(data)
}
\arguments{
\item{data}{A data.frame.}
}
\value{
\code{data} with cleaned column names.
}
\description{
Clean column names to a unique, syntactic form
}
