% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primitives.R
\name{collapselevels}
\alias{collapselevels}
\title{Collapse factor levels into named groups}
\usage{
collapselevels(x, groups)
}
\arguments{
\item{x}{An atomic vector.}

\item{groups}{Named list mapping a replacement label to the values it should collapse.}
}
\value{
A character vector with levels collapsed.
}
\description{
Collapse factor levels into named groups
}
