% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primitives.R
\name{collapsetext}
\alias{collapsetext}
\title{Collapse a vector into a single string}
\usage{
collapsetext(x, sep = "")
}
\arguments{
\item{x}{An atomic vector.}

\item{sep}{Separator string.}
}
\value{
A single character string.
}
\description{
Collapse a vector into a single string
}
