% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primitives.R
\name{collapsevalues}
\alias{collapsevalues}
\title{Collapse values into named groups}
\usage{
collapsevalues(x, groups)
}
\arguments{
\item{x}{An atomic vector.}

\item{groups}{Named list mapping a replacement label to the values it should collapse.}
}
\value{
\code{x} with values replaced by their group label.
}
\description{
Collapse values into named groups
}
