% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primitives.R
\name{colnames}
\alias{colnames}
\title{Column names}
\usage{
colnames(data)
}
\arguments{
\item{data}{A data.frame or data.table.}
}
\value{
A character vector of column names.
}
\description{
Column names
}
