% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primitives.R
\name{constants}
\alias{constants}
\title{Columns with a single distinct non-missing value}
\usage{
constants(data)
}
\arguments{
\item{data}{A data.frame.}
}
\value{
A character vector of constant column names.
}
\description{
Columns with a single distinct non-missing value
}
