\name{contains}
\alias{contains}
\title{Detect literal substrings}
\usage{
contains(x, pattern, ignore_case = FALSE)
}
\arguments{
\item{x}{A character vector or factor.}

\item{pattern}{A literal substring to search for.}

\item{ignore_case}{Whether case should be ignored.}
}
\value{
A logical vector indicating whether each element contains the pattern.
}
\description{
Detect whether text contains a literal substring.
}
