% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primitives.R
\name{countmatch}
\alias{countmatch}
\title{Count pattern matches per string}
\usage{
countmatch(x, pattern, fixed = FALSE)
}
\arguments{
\item{x}{An atomic vector.}

\item{pattern}{A regular expression (or fixed string when \code{fixed = TRUE}).}

\item{fixed}{Use fixed (non-regex) matching instead of regular expressions.}
}
\value{
An integer vector.
}
\description{
Count pattern matches per string
}
