% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primitives.R
\name{cumcount}
\alias{cumcount}
\title{Cumulative row counter}
\usage{
cumcount(x)
}
\arguments{
\item{x}{An atomic vector.}
}
\value{
An integer sequence along \code{x}.
}
\description{
Cumulative row counter
}
