% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primitives.R
\name{cumedist}
\alias{cumedist}
\title{Cumulative empirical distribution}
\usage{
cumedist(x)
}
\arguments{
\item{x}{An atomic vector.}
}
\value{
A numeric vector.
}
\description{
Cumulative empirical distribution
}
