% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primitives.R
\name{datediff}
\alias{datediff}
\title{Difference between two dates}
\usage{
datediff(x, y, units = "days")
}
\arguments{
\item{x}{An atomic vector.}

\item{y}{An atomic vector, data.frame, or basetable, depending on the function.}

\item{units}{Unit used to express the difference.}
}
\value{
A numeric vector of differences.
}
\description{
Difference between two dates
}
