% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primitives.R
\name{dateseq}
\alias{dateseq}
\title{Sequence of dates}
\usage{
dateseq(from, to, by = "day", length.out = NULL)
}
\arguments{
\item{from}{Start date.}

\item{to}{End date, or target range for rescaling.}

\item{by}{Character vector of column names identifying groups or join keys.}

\item{length.out}{Desired sequence length.}
}
\value{
A Date vector.
}
\description{
Sequence of dates
}
