% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primitives.R
\name{denserank}
\alias{denserank}
\title{Dense rank of a vector}
\usage{
denserank(x)
}
\arguments{
\item{x}{An atomic vector.}
}
\value{
An integer vector of dense ranks.
}
\description{
Dense rank of a vector
}
