% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primitives.R
\name{difference}
\alias{difference}
\title{Lagged difference}
\usage{
difference(x, lag = 1L)
}
\arguments{
\item{x}{An atomic vector.}

\item{lag}{Lag used for the difference.}
}
\value{
A numeric vector of differences.
}
\description{
Lagged difference
}
