% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primitives.R
\name{duplicatekeys}
\alias{duplicatekeys}
\title{Duplicated key combinations}
\usage{
duplicatekeys(data, by)
}
\arguments{
\item{data}{A data.frame.}

\item{by}{Character vector of column names identifying groups or join keys.}
}
\value{
See \code{\link[=duplicated_keys]{duplicated_keys()}}.
}
\description{
Duplicated key combinations
}
