% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primitives.R
\name{duplicatenames}
\alias{duplicatenames}
\title{Duplicated column names}
\usage{
duplicatenames(data)
}
\arguments{
\item{data}{A data.frame.}
}
\value{
A character vector of the duplicated names.
}
\description{
Duplicated column names
}
