\name{endswith}
\alias{endswith}
\title{Detect string suffixes}
\usage{
endswith(x, suffix, ignore_case = FALSE)
}
\arguments{
\item{x}{A character vector or factor.}

\item{suffix}{The suffix to look for.}

\item{ignore_case}{Whether case should be ignored.}
}
\value{
A logical vector indicating whether each element ends with the suffix.
}
\description{
Detect whether text ends with a given suffix.
}
