% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primitives.R
\name{equalrows}
\alias{equalrows}
\title{Compare two tables' rows for equality}
\usage{
equalrows(x, y, by = NULL)
}
\arguments{
\item{x}{An atomic vector.}

\item{y}{An atomic vector, data.frame, or basetable, depending on the function.}

\item{by}{Character vector of column names identifying groups or join keys.}
}
\value{
A single logical value.
}
\description{
Compare two tables' rows for equality
}
