% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primitives.R
\name{expandlevels}
\alias{expandlevels}
\title{Add levels to a factor}
\usage{
expandlevels(x, levels)
}
\arguments{
\item{x}{An atomic vector.}

\item{levels}{Character vector of factor levels.}
}
\value{
A factor including the additional levels.
}
\description{
Add levels to a factor
}
