% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primitives.R
\name{extractbetween}
\alias{extractbetween}
\title{Extract text between two markers}
\usage{
extractbetween(x, left, right)
}
\arguments{
\item{x}{An atomic vector.}

\item{left}{Left marker.}

\item{right}{Right marker.}
}
\value{
A character vector.
}
\description{
Extract text between two markers
}
