% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primitives.R
\name{extractint}
\alias{extractint}
\title{Extract an integer value from text}
\usage{
extractint(x)
}
\arguments{
\item{x}{An atomic vector.}
}
\value{
An integer vector.
}
\description{
Extract an integer value from text
}
