% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primitives.R
\name{extractnum}
\alias{extractnum}
\title{Extract a numeric value from text}
\usage{
extractnum(x)
}
\arguments{
\item{x}{An atomic vector.}
}
\value{
A numeric vector.
}
\description{
Extract a numeric value from text
}
