% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verbs.R
\name{filter}
\alias{filter}
\title{Filter rows}
\usage{
filter(data, ...)
}
\arguments{
\item{data}{A data frame or data.table.}

\item{...}{Logical expressions evaluated in \code{data}.}
}
\value{
A data.table.
}
\description{
Keep rows where all supplied logical expressions evaluate to \code{TRUE}.
}
