% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primitives.R
\name{firstcols}
\alias{firstcols}
\title{Move columns to the front}
\usage{
firstcols(data, cols)
}
\arguments{
\item{data}{A data.frame.}

\item{cols}{Character vector of column names.}
}
\value{
\code{data} with \code{cols} moved first.
}
\description{
Move columns to the front
}
