% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primitives.R
\name{floordate}
\alias{floordate}
\title{Round a date down to a unit}
\usage{
floordate(x, unit = c("day", "week", "month", "year"))
}
\arguments{
\item{x}{An atomic vector.}

\item{unit}{Time unit to round to.}
}
\value{
A Date vector.
}
\description{
Round a date down to a unit
}
