% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primitives.R
\name{invalidrows}
\alias{invalidrows}
\title{Rows that fail a condition}
\usage{
invalidrows(data, condition)
}
\arguments{
\item{data}{A data.frame.}

\item{condition}{A logical expression evaluated in the context of \code{data}.}
}
\value{
The rows of \code{data} for which \code{condition} is not \code{TRUE}.
}
\description{
Rows that fail a condition
}
