% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primitives.R
\name{invalidvalues}
\alias{invalidvalues}
\title{Rows whose value is not in the allowed set}
\usage{
invalidvalues(data, column, allowed)
}
\arguments{
\item{data}{A data.frame.}

\item{column}{Name of a single column.}

\item{allowed}{Vector of permitted values.}
}
\value{
The rows of \code{data} with disallowed values.
}
\description{
Rows whose value is not in the allowed set
}
