% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basetable-class.R
\name{is_basetable}
\alias{is_basetable}
\title{Test whether an object is a basetable}
\usage{
is_basetable(x)
}
\arguments{
\item{x}{An object.}
}
\value{
A single logical.
}
\description{
Test whether an object is a basetable
}
\examples{
is_basetable(as_basetable(mtcars))
is_basetable(mtcars)
}
\seealso{
\code{\link[=as_basetable]{as_basetable()}}
}
