% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primitives.R
\name{isalpha}
\alias{isalpha}
\title{Test for alphabetic-only strings}
\usage{
isalpha(x)
}
\arguments{
\item{x}{An atomic vector.}
}
\value{
A logical vector.
}
\description{
Test for alphabetic-only strings
}
