\name{isalphanumeric}
\alias{isalphanumeric}
\title{Detect alphanumeric text}
\usage{
isalphanumeric(x)
}
\arguments{
\item{x}{A character vector or factor.}
}
\value{
A logical vector indicating whether each element contains only alphanumeric characters.
}
\description{
Detect whether text contains only letters and numbers.
}
