% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primitives.R
\name{isblank}
\alias{isblank}
\title{Test for blank (missing or empty) values}
\usage{
isblank(x)
}
\arguments{
\item{x}{An atomic vector.}
}
\value{
A logical vector.
}
\description{
Test for blank (missing or empty) values
}
