% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primitives.R
\name{isemail}
\alias{isemail}
\title{Test whether text looks like an email address}
\usage{
isemail(x)
}
\arguments{
\item{x}{An atomic vector.}
}
\value{
A logical vector.
}
\description{
Test whether text looks like an email address
}
