% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primitives.R
\name{isintegertext}
\alias{isintegertext}
\title{Test whether text looks like an integer}
\usage{
isintegertext(x)
}
\arguments{
\item{x}{An atomic vector.}
}
\value{
A logical vector.
}
\description{
Test whether text looks like an integer
}
