% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primitives.R
\name{isnumerictext}
\alias{isnumerictext}
\title{Test whether text looks numeric}
\usage{
isnumerictext(x)
}
\arguments{
\item{x}{An atomic vector.}
}
\value{
A logical vector.
}
\description{
Test whether text looks numeric
}
