% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primitives.R
\name{isurl}
\alias{isurl}
\title{Test whether text looks like a URL}
\usage{
isurl(x)
}
\arguments{
\item{x}{An atomic vector.}
}
\value{
A logical vector.
}
\description{
Test whether text looks like a URL
}
