% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primitives.R
\name{joinrelationship}
\alias{joinrelationship}
\title{Cardinality of a join key relationship}
\usage{
joinrelationship(x, y, by)
}
\arguments{
\item{x}{An atomic vector.}

\item{y}{An atomic vector, data.frame, or basetable, depending on the function.}

\item{by}{Character vector of column names identifying groups or join keys.}
}
\value{
One of \code{"one-to-one"}, \code{"one-to-many"}, \code{"many-to-one"}, \code{"many-to-many"}.
}
\description{
Cardinality of a join key relationship
}
