\name{jointext}
\alias{jointext}
\title{Join text elements}
\usage{
jointext(..., sep = " ")
}
\arguments{
\item{...}{Character vectors to join.}

\item{sep}{Separator placed between joined pieces.}
}
\value{
A character vector containing the joined text.
}
\description{
Join text elements using a separator.
}
