% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primitives.R
\name{lagvalue}
\alias{lagvalue}
\title{Lag a vector}
\usage{
lagvalue(x, n = 1L, default = NA)
}
\arguments{
\item{x}{An atomic vector.}

\item{n}{Integer count.}

\item{default}{Value used to fill positions with no lagged observation.}
}
\value{
\code{x} shifted forward by \code{n} positions.
}
\description{
Lag a vector
}
