% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primitives.R
\name{lastrows}
\alias{lastrows}
\title{Last \code{n} rows}
\usage{
lastrows(data, n = 1L)
}
\arguments{
\item{data}{A data.frame.}

\item{n}{Integer count.}
}
\value{
The last \code{n} rows of \code{data}.
}
\description{
Last \code{n} rows
}
