% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primitives.R
\name{leadvalue}
\alias{leadvalue}
\title{Lead a vector}
\usage{
leadvalue(x, n = 1L, default = NA)
}
\arguments{
\item{x}{An atomic vector.}

\item{n}{Integer count.}

\item{default}{Value used to fill positions with no leading observation.}
}
\value{
\code{x} shifted backward by \code{n} positions.
}
\description{
Lead a vector
}
