\name{left}
\alias{left}
\title{Take leading characters}
\usage{
left(x, n = 1L)
}
\arguments{
\item{x}{A character vector or factor.}

\item{n}{Number of leading characters to keep.}
}
\value{
A character vector containing the leading characters.
}
\description{
Take the leftmost characters from each element of text.
}
