% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primitives.R
\name{locateall}
\alias{locateall}
\title{Positions of all pattern matches}
\usage{
locateall(x, pattern, fixed = FALSE)
}
\arguments{
\item{x}{An atomic vector.}

\item{pattern}{A regular expression (or fixed string when \code{fixed = TRUE}).}

\item{fixed}{Use fixed (non-regex) matching instead of regular expressions.}
}
\value{
A list of match positions (see \code{\link[=gregexpr]{gregexpr()}}).
}
\description{
Positions of all pattern matches
}
