% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primitives.R
\name{lump}
\alias{lump}
\title{Lump infrequent values into "Other"}
\usage{
lump(x, n = 5, other = "Other")
}
\arguments{
\item{x}{An atomic vector.}

\item{n}{Integer count.}

\item{other}{Label used for values outside the top \code{n}.}
}
\value{
A character vector.
}
\description{
Lump infrequent values into "Other"
}
