% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primitives.R
\name{matchedkeys}
\alias{matchedkeys}
\title{Rows of \code{x} whose key is present in \code{y}}
\usage{
matchedkeys(x, y, by)
}
\arguments{
\item{x}{An atomic vector.}

\item{y}{An atomic vector, data.frame, or basetable, depending on the function.}

\item{by}{Character vector of column names identifying groups or join keys.}
}
\value{
Rows of \code{x} whose key is present in \code{y}.
}
\description{
Rows of \code{x} whose key is present in \code{y}
}
