\name{middle}
\alias{middle}
\title{Take characters from the middle}
\usage{
middle(x, start, end)
}
\arguments{
\item{x}{A character vector or factor.}

\item{start}{Starting character position.}

\item{end}{Ending character position.}
}
\value{
A character vector containing the selected substring.
}
\description{
Take a substring from each element using explicit start and end positions.
}
