% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missing.R
\name{missingness}
\alias{missingness}
\title{Summarize missing values by column or row}
\usage{
missingness(data, margin = c("column", "row"))
}
\arguments{
\item{data}{A data.frame.}

\item{margin}{Summarize missingness \code{"column"}-wise (the default) or
\code{"row"}-wise.}
}
\value{
A basetable with one row per column (or per row) describing the count
and proportion of missing values.
}
\description{
Summarize missing values by column or row
}
