% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primitives.R
\name{move}
\alias{move}
\title{Move columns before or after another column}
\usage{
move(data, cols, before = NULL, after = NULL)
}
\arguments{
\item{data}{A data.frame.}

\item{cols}{Character vector of column names.}

\item{before}{Column identifying where to insert \code{cols} before.}

\item{after}{Column identifying where to insert \code{cols} after.}
}
\value{
\code{data} with columns reordered.
}
\description{
Move columns before or after another column
}
