% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primitives.R
\name{naif}
\alias{naif}
\title{Recode matching values to \code{NA}}
\usage{
naif(x, values)
}
\arguments{
\item{x}{An atomic vector.}

\item{values}{Vector or list of replacement values.}
}
\value{
\code{x} with matches replaced by \code{NA}.
}
\description{
Recode matching values to \code{NA}
}
