% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primitives.R
\name{nato}
\alias{nato}
\title{Replace \code{NA} with a value}
\usage{
nato(x, value)
}
\arguments{
\item{x}{An atomic vector.}

\item{value}{Value to test, assign, or match against.}
}
\value{
\code{x} with missing values replaced.
}
\description{
Replace \code{NA} with a value
}
