% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primitives.R
\name{natoblank}
\alias{natoblank}
\title{Replace \code{NA} with an empty string}
\usage{
natoblank(x)
}
\arguments{
\item{x}{An atomic vector.}
}
\value{
\code{x} with missing values replaced by \code{""}.
}
\description{
Replace \code{NA} with an empty string
}
