% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ncols.R
\name{ncols}
\alias{ncols}
\title{Number of columns}
\usage{
ncols(data)
}
\arguments{
\item{data}{A data frame or data table.}
}
\value{
An integer scalar.
}
\description{
Return the number of columns in a table.
}
