% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primitives.R
\name{nearestmerge}
\alias{nearestmerge}
\title{Nearest-key join}
\usage{
nearestmerge(x, y, by, tolerance = Inf)
}
\arguments{
\item{x}{An atomic vector.}

\item{y}{An atomic vector, data.frame, or basetable, depending on the function.}

\item{by}{Character vector of column names identifying groups or join keys.}

\item{tolerance}{Numeric join/comparison tolerance.}
}
\value{
See \code{\link[=rollingmerge]{rollingmerge()}}.
}
\description{
Nearest-key join
}
