% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primitives.R
\name{normalizeencoding}
\alias{normalizeencoding}
\title{Normalize text encoding to UTF-8}
\usage{
normalizeencoding(x)
}
\arguments{
\item{x}{An atomic vector.}
}
\value{
A character vector.
}
\description{
Normalize text encoding to UTF-8
}
