% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primitives.R
\name{orderrows}
\alias{orderrows}
\title{Order rows by one or more columns}
\usage{
orderrows(data, by, decreasing = FALSE, na.last = TRUE)
}
\arguments{
\item{data}{A data.frame.}

\item{by}{Character vector of column names identifying groups or join keys.}

\item{decreasing}{Sort in decreasing order.}

\item{na.last}{Placement of missing values when sorting.}
}
\value{
\code{data} sorted by \code{by}.
}
\description{
Order rows by one or more columns
}
