% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primitives.R
\name{padleft}
\alias{padleft}
\title{Pad text on the left}
\usage{
padleft(x, width, pad = " ")
}
\arguments{
\item{x}{An atomic vector.}

\item{width}{Target width, in characters or rolling-window size depending on the function.}

\item{pad}{Padding character.}
}
\value{
A character vector padded to \code{width}.
}
\description{
Pad text on the left
}
