% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primitives.R
\name{parsecurrency}
\alias{parsecurrency}
\title{Parse a currency string as a number}
\usage{
parsecurrency(x, strict = FALSE)
}
\arguments{
\item{x}{An atomic vector.}

\item{strict}{Raise an error if any non-missing value fails to parse.}
}
\value{
A numeric vector.
}
\description{
Parse a currency string as a number
}
