% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primitives.R
\name{parsedate}
\alias{parsedate}
\title{Parse text as dates, trying multiple formats}
\usage{
parsedate(x, formats, tz = "UTC", strict = FALSE)
}
\arguments{
\item{x}{An atomic vector.}

\item{formats}{Candidate format strings tried in order.}

\item{tz}{Time zone used when parsing.}

\item{strict}{Raise an error if any non-missing value fails to parse.}
}
\value{
A Date vector.
}
\description{
Parse text as dates, trying multiple formats
}
