% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primitives.R
\name{parsedatetime}
\alias{parsedatetime}
\title{Parse text as date-times, trying multiple formats}
\usage{
parsedatetime(x, formats, tz = "UTC", strict = FALSE)
}
\arguments{
\item{x}{An atomic vector.}

\item{formats}{Candidate format strings tried in order.}

\item{tz}{Time zone used when parsing.}

\item{strict}{Raise an error if any non-missing value fails to parse.}
}
\value{
A POSIXct vector.
}
\description{
Parse text as date-times, trying multiple formats
}
