% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primitives.R
\name{parsefailures}
\alias{parsefailures}
\title{Rows where parsing failed}
\usage{
parsefailures(x, fun, ...)
}
\arguments{
\item{x}{An atomic vector.}

\item{fun}{Function applied to each element, column, or group.}

\item{...}{Additional arguments (unused, or passed through depending on the function).}
}
\value{
A basetable of the failed indices and values.
}
\description{
Rows where parsing failed
}
