% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primitives.R
\name{parseint}
\alias{parseint}
\title{Parse text as integers}
\usage{
parseint(x, na = character(), strict = FALSE)
}
\arguments{
\item{x}{An atomic vector.}

\item{na}{Values treated as missing before parsing.}

\item{strict}{Raise an error if any non-missing value fails to parse.}
}
\value{
An integer vector.
}
\description{
Parse text as integers
}
