% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primitives.R
\name{parselogical}
\alias{parselogical}
\title{Parse text as logicals}
\usage{
parselogical(x, na = character(), strict = FALSE)
}
\arguments{
\item{x}{An atomic vector.}

\item{na}{Values treated as missing before parsing.}

\item{strict}{Raise an error if any non-missing value fails to parse.}
}
\value{
A logical vector.
}
\description{
Parse text as logicals
}
