% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primitives.R
\name{parsepercent}
\alias{parsepercent}
\title{Parse a percentage string as a number}
\usage{
parsepercent(x, strict = FALSE)
}
\arguments{
\item{x}{An atomic vector.}

\item{strict}{Raise an error if any non-missing value fails to parse.}
}
\value{
A numeric vector.
}
\description{
Parse a percentage string as a number
}
