% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primitives.R
\name{percentchange}
\alias{percentchange}
\title{Period-over-period percent change}
\usage{
percentchange(x)
}
\arguments{
\item{x}{An atomic vector.}
}
\value{
A numeric vector of percent changes.
}
\description{
Period-over-period percent change
}
