% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primitives.R
\name{percentrank}
\alias{percentrank}
\title{Percentile rank of a vector}
\usage{
percentrank(x)
}
\arguments{
\item{x}{An atomic vector.}
}
\value{
A numeric vector of percentile ranks in \verb{[0, 1]}.
}
\description{
Percentile rank of a vector
}
