% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primitives.R
\name{quantilegroup}
\alias{quantilegroup}
\title{Bin a vector into quantile groups}
\usage{
quantilegroup(x, n = 5)
}
\arguments{
\item{x}{An atomic vector.}

\item{n}{Integer count.}
}
\value{
A factor of quantile bins.
}
\description{
Bin a vector into quantile groups
}
