% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primitives.R
\name{recombine}
\alias{recombine}
\title{Recombine a list of pieces into one table}
\usage{
recombine(x, id = NULL)
}
\arguments{
\item{x}{An atomic vector.}

\item{id}{Optional name for a source identifier column.}
}
\value{
See \code{\link[=combine]{combine()}}.
}
\description{
Recombine a list of pieces into one table
}
