% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primitives.R
\name{removeall}
\alias{removeall}
\title{Remove text matching a pattern (alias)}
\usage{
removeall(x, pattern, fixed = FALSE)
}
\arguments{
\item{x}{An atomic vector.}

\item{pattern}{A regular expression (or fixed string when \code{fixed = TRUE}).}

\item{fixed}{Use fixed (non-regex) matching instead of regular expressions.}
}
\value{
A character vector.
}
\description{
Remove text matching a pattern (alias)
}
