% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primitives.R
\name{reorderlevels}
\alias{reorderlevels}
\title{Reorder factor levels}
\usage{
reorderlevels(x, by)
}
\arguments{
\item{x}{An atomic vector.}

\item{by}{Character vector of column names identifying groups or join keys.}
}
\value{
A factor with reordered levels.
}
\description{
Reorder factor levels
}
