% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primitives.R
\name{repairnames}
\alias{repairnames}
\title{Repair column names}
\usage{
repairnames(data, method = c("unique", "universal", "minimal"))
}
\arguments{
\item{data}{A data.frame.}

\item{method}{Cleaning strategy to apply to names.}
}
\value{
\code{data} with repaired column names.
}
\description{
Repair column names
}
