% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primitives.R
\name{replaceall}
\alias{replaceall}
\title{Recode values by lookup (alias)}
\usage{
replaceall(x, old, new)
}
\arguments{
\item{x}{An atomic vector.}

\item{old}{Existing value(s) to replace.}

\item{new}{Replacement value(s) matching \code{old} by position.}
}
\value{
See \code{\link[=replacevalues]{replacevalues()}}.
}
\description{
Recode values by lookup (alias)
}
