% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primitives.R
\name{replacetext}
\alias{replacetext}
\title{Replace a pattern with replacement text}
\usage{
replacetext(x, pattern, replacement, fixed = FALSE)
}
\arguments{
\item{x}{An atomic vector.}

\item{pattern}{A regular expression (or fixed string when \code{fixed = TRUE}).}

\item{replacement}{Replacement text.}

\item{fixed}{Use fixed (non-regex) matching instead of regular expressions.}
}
\value{
A character vector.
}
\description{
Replace a pattern with replacement text
}
