% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primitives.R
\name{replacevalues}
\alias{replacevalues}
\title{Recode values by lookup}
\usage{
replacevalues(x, old, new)
}
\arguments{
\item{x}{An atomic vector.}

\item{old}{Baseline data.frame ("before" state).}

\item{new}{Updated data.frame ("after" state), or replacement values when used for value substitution.}
}
\value{
\code{x} with matched values replaced.
}
\description{
Recode values by lookup
}
