% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primitives.R
\name{rescale}
\alias{rescale}
\title{Rescale a vector to a new range}
\usage{
rescale(x, to = c(0, 1))
}
\arguments{
\item{x}{An atomic vector.}

\item{to}{End date, or target range for rescaling.}
}
\value{
A rescaled numeric vector.
}
\description{
Rescale a vector to a new range
}
