% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primitives.R
\name{reverse}
\alias{reverse}
\title{Reverse row order}
\usage{
reverse(data)
}
\arguments{
\item{data}{A data.frame or data.table.}
}
\value{
\code{data} with rows in reverse order.
}
\description{
Reverse row order
}
