\name{right}
\alias{right}
\title{Take trailing characters}
\usage{
right(x, n = 1L)
}
\arguments{
\item{x}{A character vector or factor.}

\item{n}{Number of trailing characters to keep.}
}
\value{
A character vector containing the trailing characters.
}
\description{
Take the rightmost characters from each element of text.
}
