% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primitives.R
\name{rounddate}
\alias{rounddate}
\title{Round a date to the nearest unit}
\usage{
rounddate(x, unit = c("day", "week", "month", "year"))
}
\arguments{
\item{x}{An atomic vector.}

\item{unit}{Time unit to round to.}
}
\value{
A Date vector.
}
\description{
Round a date to the nearest unit
}
