% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primitives.R
\name{samplerows}
\alias{samplerows}
\title{Sample \code{n} rows without replacement}
\usage{
samplerows(data, n)
}
\arguments{
\item{data}{A data.frame.}

\item{n}{Integer count.}
}
\value{
A random sample of \code{n} rows.
}
\description{
Sample \code{n} rows without replacement
}
