% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primitives.R
\name{sentencecase}
\alias{sentencecase}
\title{Convert to sentence case}
\usage{
sentencecase(x)
}
\arguments{
\item{x}{An atomic vector.}
}
\value{
A sentence-case character vector.
}
\description{
Convert to sentence case
}
