% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primitives.R
\name{similartext}
\alias{similartext}
\title{Nearest string match (alias)}
\usage{
similartext(x, choices)
}
\arguments{
\item{x}{An atomic vector.}

\item{choices}{Candidate strings to match against.}
}
\value{
See \code{\link[=nearesttext]{nearesttext()}}.
}
\description{
Nearest string match (alias)
}
