% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primitives.R
\name{splitlast}
\alias{splitlast}
\title{Last piece after splitting on a separator}
\usage{
splitlast(x, sep, fixed = FALSE)
}
\arguments{
\item{x}{An atomic vector.}

\item{sep}{Separator string.}

\item{fixed}{Use fixed (non-regex) matching instead of regular expressions.}
}
\value{
A character vector.
}
\description{
Last piece after splitting on a separator
}
