\name{splittext}
\alias{splittext}
\title{Split text on a literal delimiter}
\usage{
splittext(x, pattern)
}
\arguments{
\item{x}{A character vector or factor.}

\item{pattern}{A literal delimiter to split on.}
}
\value{
A list of character vectors containing the split pieces.
}
\description{
Split each element of text on a literal delimiter.
}
