% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primitives.R
\name{squish}
\alias{squish}
\title{Trim and collapse internal whitespace}
\usage{
squish(x)
}
\arguments{
\item{x}{An atomic vector.}
}
\value{
A cleaned character vector.
}
\description{
Trim and collapse internal whitespace
}
