% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primitives.R
\name{standardize}
\alias{standardize}
\title{Standardize a vector to mean 0, SD 1}
\usage{
standardize(x)
}
\arguments{
\item{x}{An atomic vector.}
}
\value{
A standardized numeric vector.
}
\description{
Standardize a vector to mean 0, SD 1
}
