\name{startswith}
\alias{startswith}
\title{Detect string prefixes}
\usage{
startswith(x, prefix, ignore_case = FALSE)
}
\arguments{
\item{x}{A character vector or factor.}

\item{prefix}{The prefix to look for.}

\item{ignore_case}{Whether case should be ignored.}
}
\value{
A logical vector indicating whether each element starts with the prefix.
}
\description{
Detect whether text starts with a given prefix.
}
