% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primitives.R
\name{textdist}
\alias{textdist}
\title{Pairwise string distances}
\usage{
textdist(x, y)
}
\arguments{
\item{x}{An atomic vector.}

\item{y}{An atomic vector, data.frame, or basetable, depending on the function.}
}
\value{
A numeric distance matrix (see \code{\link[utils:adist]{utils::adist()}}).
}
\description{
Pairwise string distances
}
