% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primitives.R
\name{titlecase}
\alias{titlecase}
\title{Convert to title case}
\usage{
titlecase(x)
}
\arguments{
\item{x}{An atomic vector.}
}
\value{
A title-case character vector.
}
\description{
Convert to title case
}
