\name{transmute}
\alias{transmute}
\title{Create and keep only new columns}
\usage{
transmute(data, ...)
}
\arguments{
\item{data}{A data frame or data table.}

\item{...}{Named expressions that create new columns.}
}
\value{
A data.table containing only the new columns.
}
\description{
Create columns and keep only the new variables.
}
