% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primitives.R
\name{transpose}
\alias{transpose}
\title{Transpose a table}
\usage{
transpose(data)
}
\arguments{
\item{data}{A data.frame or data.table.}
}
\value{
A transposed data.table.
}
\description{
Transpose a table
}
