% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map.R
\name{traverse}
\alias{traverse}
\title{Traverse a list of arguments}
\usage{
traverse(.l, .f, ...)
}
\arguments{
\item{.l}{A list of vectors or lists with a common length.}

\item{.f}{A function or function name.}

\item{...}{Additional arguments passed to \code{.f}.}
}
\value{
\code{traverse()} returns a list.
}
\description{
Call \code{.f} once per position across several equal-length vectors or lists,
passing the i-th element of each as a positional argument. The parallel-map
companion to \code{\link[=map]{map()}}.
}
\examples{
traverse(list(a = 1:2, b = 10:11), function(a, b) a + b)
}
