% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primitives.R
\name{trim}
\alias{trim}
\title{Trim leading and trailing whitespace}
\usage{
trim(x)
}
\arguments{
\item{x}{An atomic vector.}
}
\value{
A trimmed character vector.
}
\description{
Trim leading and trailing whitespace
}
