% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primitives.R
\name{truncate}
\alias{truncate}
\title{Truncate text with an ellipsis}
\usage{
truncate(x, n, ellipsis = "...")
}
\arguments{
\item{x}{An atomic vector.}

\item{n}{Integer count.}

\item{ellipsis}{String appended to truncated text.}
}
\value{
A character vector.
}
\description{
Truncate text with an ellipsis
}
