% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primitives.R
\name{winsorize}
\alias{winsorize}
\title{Winsorize a vector at given quantiles}
\usage{
winsorize(x, probs = c(0.01, 0.99))
}
\arguments{
\item{x}{An atomic vector.}

\item{probs}{Quantile probabilities.}
}
\value{
A winsorized numeric vector.
}
\description{
Winsorize a vector at given quantiles
}
