% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primitives.R
\name{assertcomplete}
\alias{assertcomplete}
\title{Assert that there are no missing values}
\usage{
assertcomplete(data, cols = NULL)
}
\arguments{
\item{data}{A data.frame or data.table.}

\item{cols}{Character vector of column names.}
}
\value{
\code{data}, invisibly, if the assertion passes.
}
\description{
Assert that there are no missing values
}
