% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primitives.R
\name{assertkey}
\alias{assertkey}
\title{Assert that a key is unique (alias)}
\usage{
assertkey(data, by)
}
\arguments{
\item{data}{A data.frame or data.table.}

\item{by}{Character vector of column names identifying groups or join keys.}
}
\value{
See \code{\link[=assert_key]{assert_key()}}.
}
\description{
Assert that a key is unique (alias)
}
