% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primitives.R
\name{assertnames}
\alias{assertnames}
\title{Assert that required column names are present}
\usage{
assertnames(data, names)
}
\arguments{
\item{data}{A data.frame or data.table.}

\item{names}{Required column names.}
}
\value{
\code{data}, invisibly, if the assertion passes.
}
\description{
Assert that required column names are present
}
