% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primitives.R
\name{asserttype}
\alias{asserttype}
\title{Assert that a column has an expected class}
\usage{
asserttype(data, column, class)
}
\arguments{
\item{data}{A data.frame or data.table.}

\item{column}{Name of a single column.}

\item{class}{Class name to check against.}
}
\value{
\code{data}, invisibly, if the assertion passes.
}
\description{
Assert that a column has an expected class
}
