% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primitives.R
\name{assertunique}
\alias{assertunique}
\title{Assert that selected columns are unique}
\usage{
assertunique(data, cols)
}
\arguments{
\item{data}{A data.frame or data.table.}

\item{cols}{Character vector of column names.}
}
\value{
\code{data}, invisibly, if the assertion passes.
}
\description{
Assert that selected columns are unique
}
