% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primitives.R
\name{assertvalues}
\alias{assertvalues}
\title{Assert that a column only contains allowed values}
\usage{
assertvalues(data, column, allowed)
}
\arguments{
\item{data}{A data.frame or data.table.}

\item{column}{Name of a single column.}

\item{allowed}{Vector of permitted values.}
}
\value{
\code{data}, invisibly, if the assertion passes.
}
\description{
Assert that a column only contains allowed values
}
