% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primitives.R
\name{classes}
\alias{classes}
\title{Column classes}
\usage{
classes(data)
}
\arguments{
\item{data}{A data.frame or data.table.}
}
\value{
A tibble with one row per column giving its class.
}
\description{
Column classes
}
