\name{collapsetext}
\alias{collapsetext}
\title{Collapse text into one string}
\usage{
collapsetext(x, sep = ", ")
}
\arguments{
\item{x}{A character vector or factor.}

\item{sep}{Separator placed between collapsed pieces.}
}
\value{
A single character string.
}
\description{
Collapse text elements into one string using a separator.
}
