% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primitives.R
\name{commonnames}
\alias{commonnames}
\title{Column names shared by two tables}
\usage{
commonnames(x, y)
}
\arguments{
\item{x}{An atomic vector.}

\item{y}{An atomic vector, data.frame, or data.table, depending on the function.}
}
\value{
See \code{\link[=common_names]{common_names()}}.
}
\description{
Column names shared by two tables
}
