% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primitives.R
\name{compareschema}
\alias{compareschema}
\title{Compare the schemas of two tables}
\usage{
compareschema(x, y)
}
\arguments{
\item{x}{An atomic vector.}

\item{y}{An atomic vector, data.frame, or data.table, depending on the function.}
}
\value{
A tibble describing shared and differing columns/types.
}
\description{
Compare the schemas of two tables
}
