% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primitives.R
\name{duplicaterows}
\alias{duplicaterows}
\title{Rows involved in a duplicate}
\usage{
duplicaterows(data)
}
\arguments{
\item{data}{A data.frame or data.table.}
}
\value{
A data.frame of the duplicated rows.
}
\description{
Rows involved in a duplicate
}
