% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primitives.R
\name{emptycols}
\alias{emptycols}
\title{Columns that are entirely blank or missing}
\usage{
emptycols(data)
}
\arguments{
\item{data}{A data.frame or data.table.}
}
\value{
A character vector of column names.
}
\description{
Columns that are entirely blank or missing
}
