% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primitives.R
\name{emptyrows}
\alias{emptyrows}
\title{Rows that are entirely blank or missing}
\usage{
emptyrows(data)
}
\arguments{
\item{data}{A data.frame or data.table.}
}
\value{
A data.frame of the fully blank rows.
}
\description{
Rows that are entirely blank or missing
}
