% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primitives.R
\name{endswith}
\alias{endswith}
\title{Test whether strings end with a pattern}
\usage{
endswith(x, pattern, fixed = FALSE)
}
\arguments{
\item{x}{An atomic vector.}

\item{pattern}{A regular expression (or fixed string when \code{fixed = TRUE}).}

\item{fixed}{Use fixed (non-regex) matching instead of regular expressions.}
}
\value{
A logical vector.
}
\description{
Test whether strings end with a pattern
}
