% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primitives.R
\name{equaldata}
\alias{equaldata}
\title{Compare two tables for equality}
\usage{
equaldata(
  x,
  y,
  ignoreorder = FALSE,
  ignorerownames = TRUE,
  tolerance = sqrt(.Machine$double.eps)
)
}
\arguments{
\item{x}{An atomic vector.}

\item{y}{An atomic vector, data.frame, or data.table, depending on the function.}

\item{ignoreorder}{Ignore row order when comparing.}

\item{ignorerownames}{Ignore row names when comparing.}

\item{tolerance}{Numeric join/comparison tolerance.}
}
\value{
A single logical value.
}
\description{
Compare two tables for equality
}
