% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primitives.R
\name{firstrows}
\alias{firstrows}
\title{First \code{n} rows}
\usage{
firstrows(data, n = 1L)
}
\arguments{
\item{data}{A data.frame or data.table.}

\item{n}{Integer count.}
}
\value{
The first \code{n} rows of \code{data}.
}
\description{
First \code{n} rows
}
