% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primitives.R
\name{intersectrows}
\alias{intersectrows}
\title{Set intersection of rows}
\usage{
intersectrows(x, y, by = NULL)
}
\arguments{
\item{x}{An atomic vector.}

\item{y}{An atomic vector, data.frame, or basetable, depending on the function.}

\item{by}{Character vector of column names identifying groups or join keys.}
}
\value{
Rows of \code{x} also present in \code{y}.
}
\description{
Set intersection of rows
}
