% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primitives.R
\name{isalphanumeric}
\alias{isalphanumeric}
\title{Test for alphanumeric-only strings}
\usage{
isalphanumeric(x)
}
\arguments{
\item{x}{An atomic vector.}
}
\value{
A logical vector.
}
\description{
Test for alphanumeric-only strings
}
