\name{isemail}
\alias{isemail}
\title{Detect email-like text}
\usage{
isemail(x)
}
\arguments{
\item{x}{A character vector or factor.}
}
\value{
A logical vector indicating whether each element looks like an email address.
}
\description{
Detect whether text looks like an email address using an explicit regular expression.
}
