\name{isintegertext}
\alias{isintegertext}
\title{Detect integer text}
\usage{
isintegertext(x)
}
\arguments{
\item{x}{A character vector or factor.}
}
\value{
A logical vector indicating whether each element looks like an integer string.
}
\description{
Detect whether text looks like an integer value using an explicit regular expression.
}
