\name{isnumerictext}
\alias{isnumerictext}
\title{Detect numeric text}
\usage{
isnumerictext(x)
}
\arguments{
\item{x}{A character vector or factor.}
}
\value{
A logical vector indicating whether each element looks like a numeric string.
}
\description{
Detect whether text looks like a numeric value using an explicit regular expression.
}
