% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primitives.R
\name{jointext}
\alias{jointext}
\title{Concatenate values element-wise}
\usage{
jointext(...)
}
\arguments{
\item{...}{Additional arguments (unused, or passed through depending on the function).}
}
\value{
A character vector.
}
\description{
Concatenate values element-wise
}
