% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primitives.R
\name{lastcols}
\alias{lastcols}
\title{Move columns to the back}
\usage{
lastcols(data, cols)
}
\arguments{
\item{data}{A data.frame or data.table.}

\item{cols}{Character vector of column names.}
}
\value{
\code{data} with \code{cols} moved last.
}
\description{
Move columns to the back
}
