% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primitives.R
\name{left}
\alias{left}
\title{First \code{n} characters}
\usage{
left(x, n)
}
\arguments{
\item{x}{An atomic vector.}

\item{n}{Integer count.}
}
\value{
A character vector.
}
\description{
First \code{n} characters
}
