% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primitives.R
\name{locate}
\alias{locate}
\title{Position of the first pattern match}
\usage{
locate(x, pattern, fixed = FALSE)
}
\arguments{
\item{x}{An atomic vector.}

\item{pattern}{A regular expression (or fixed string when \code{fixed = TRUE}).}

\item{fixed}{Use fixed (non-regex) matching instead of regular expressions.}
}
\value{
An integer vector of match positions (see \code{\link[=regexpr]{regexpr()}}).
}
\description{
Position of the first pattern match
}
