% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primitives.R
\name{matches}
\alias{matches}
\title{Test for a full-string pattern match}
\usage{
matches(x, pattern, fixed = FALSE)
}
\arguments{
\item{x}{An atomic vector.}

\item{pattern}{A regular expression (or fixed string when \code{fixed = TRUE}).}

\item{fixed}{Use fixed (non-regex) matching instead of regular expressions.}
}
\value{
A logical vector.
}
\description{
Test for a full-string pattern match
}
