% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primitives.R
\name{middle}
\alias{middle}
\title{Substring between two positions}
\usage{
middle(x, start, end)
}
\arguments{
\item{x}{An atomic vector.}

\item{start}{Start bound (inclusive).}

\item{end}{End bound (inclusive).}
}
\value{
A character vector.
}
\description{
Substring between two positions
}
