% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primitives.R
\name{nearesttext}
\alias{nearesttext}
\title{Nearest string match}
\usage{
nearesttext(x, choices)
}
\arguments{
\item{x}{An atomic vector.}

\item{choices}{Candidate strings to match against.}
}
\value{
A character vector of nearest matches.
}
\description{
Nearest string match
}
