% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primitives.R
\name{normalizeunicode}
\alias{normalizeunicode}
\title{Normalize Unicode text}
\usage{
normalizeunicode(x)
}
\arguments{
\item{x}{An atomic vector.}
}
\value{
A character vector (currently returned unchanged).
}
\description{
Normalize Unicode text
}
