% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primitives.R
\name{recode}
\alias{recode}
\title{Recode values by lookup (alias)}
\usage{
recode(x, old, new)
}
\arguments{
\item{x}{An atomic vector.}

\item{old}{Baseline data.frame or data.table ("before" state).}

\item{new}{Updated data.frame or data.table ("after" state), or replacement values when used for value substitution.}
}
\value{
See \code{\link[=replacevalues]{replacevalues()}}.
}
\description{
Recode values by lookup (alias)
}
