% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map.R
\name{reduce}
\alias{reduce}
\title{Reduce a vector or list}
\usage{
reduce(
  .x,
  .f,
  .init = NULL,
  .right = FALSE,
  .accumulate = FALSE,
  .simplify = TRUE
)
}
\arguments{
\item{.x}{A vector or list.}

\item{.f}{A two-argument reducing function.}

\item{.init}{Optional initial value.}

\item{.right}{Reduce from the right.}

\item{.accumulate}{Return intermediate accumulated values.}

\item{.simplify}{Simplify accumulated results when possible.}
}
\value{
The reduced value, or accumulated values when \code{.accumulate = TRUE}.
}
\description{
Reduce a vector or list using \code{functionals::freduce()}.
}
