% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verbs.R
\name{relocate}
\alias{relocate}
\title{Relocate columns}
\usage{
relocate(data, cols, .before = NULL, .after = NULL)
}
\arguments{
\item{data}{A data frame or data.table.}

\item{cols}{Character vector of column names.}

\item{.before, .after}{Optional single column name controlling placement.}
}
\value{
A tibble.
}
\description{
Move selected columns before or after another column.
}
