\name{removeaccents}
\alias{removeaccents}
\title{Remove accents from text}
\usage{
removeaccents(x)
}
\arguments{
\item{x}{A character vector or factor.}
}
\value{
A character vector with accented characters transliterated to ASCII where possible.
}
\description{
Remove accents from text using base transliteration.
}
