% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primitives.R
\name{removeaccents}
\alias{removeaccents}
\title{Strip accents from text}
\usage{
removeaccents(x)
}
\arguments{
\item{x}{An atomic vector.}
}
\value{
A character vector.
}
\description{
Strip accents from text
}
