% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primitives.R
\name{renamewith}
\alias{renamewith}
\title{Rename columns with a function}
\usage{
renamewith(data, cols, fun)
}
\arguments{
\item{data}{A data.frame or data.table.}

\item{cols}{Character vector of column names.}

\item{fun}{Function applied to each element, column, or group.}
}
\value{
\code{data} with the selected columns renamed.
}
\description{
Rename columns with a function
}
