% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primitives.R
\name{replacecols}
\alias{replacecols}
\title{Replace selected columns with new values}
\usage{
replacecols(data, cols, values)
}
\arguments{
\item{data}{A data.frame or data.table.}

\item{cols}{Character vector of column names.}

\item{values}{Vector or list of replacement values.}
}
\value{
\code{data} with the selected columns replaced.
}
\description{
Replace selected columns with new values
}
