% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primitives.R
\name{right}
\alias{right}
\title{Last \code{n} characters}
\usage{
right(x, n)
}
\arguments{
\item{x}{An atomic vector.}

\item{n}{Integer count.}
}
\value{
A character vector.
}
\description{
Last \code{n} characters
}
