% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primitives.R
\name{rowany}
\alias{rowany}
\title{Row-wise \code{any()} across columns}
\usage{
rowany(data, cols = NULL, na.rm = FALSE)
}
\arguments{
\item{data}{A data.frame or data.table.}

\item{cols}{Character vector of column names.}

\item{na.rm}{Drop missing values before computing the result.}
}
\value{
A logical vector.
}
\description{
Row-wise \code{any()} across columns
}
