% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primitives.R
\name{rowlast}
\alias{rowlast}
\title{Last non-missing value across columns, per row}
\usage{
rowlast(data, cols = NULL, na.rm = FALSE)
}
\arguments{
\item{data}{A data.frame or data.table.}

\item{cols}{Character vector of column names.}

\item{na.rm}{Drop missing values before computing the result.}
}
\value{
A vector of last values.
}
\description{
Last non-missing value across columns, per row
}
