% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primitives.R
\name{sameschema}
\alias{sameschema}
\title{Test whether two tables share the same column names}
\usage{
sameschema(x, y)
}
\arguments{
\item{x}{An atomic vector.}

\item{y}{An atomic vector, data.frame, or data.table, depending on the function.}
}
\value{
A single logical value.
}
\description{
Test whether two tables share the same column names
}
