% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primitives.R
\name{samplefrac}
\alias{samplefrac}
\title{Sample a fraction of rows without replacement}
\usage{
samplefrac(data, frac)
}
\arguments{
\item{data}{A data.frame.}

\item{frac}{Fraction of rows to sample.}
}
\value{
A random sample of rows.
}
\description{
Sample a fraction of rows without replacement
}
