% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verbs.R
\name{slice}
\alias{slice}
\title{Slice rows}
\usage{
slice(data, rows)
}
\arguments{
\item{data}{A data frame or data.table.}

\item{rows}{Integer row positions.}
}
\value{
A tibble.
}
\description{
Select rows by integer position.
}
