\name{splitfirst}
\alias{splitfirst}
\title{Take text before the first delimiter}
\usage{
splitfirst(x, pattern)
}
\arguments{
\item{x}{A character vector or factor.}

\item{pattern}{A literal delimiter.}
}
\value{
A character vector containing the first split piece from each element.
}
\description{
Return the text before the first occurrence of a literal delimiter.
}
