% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primitives.R
\name{splittext}
\alias{splittext}
\title{Split text on a separator}
\usage{
splittext(x, sep, fixed = FALSE)
}
\arguments{
\item{x}{An atomic vector.}

\item{sep}{Separator string.}

\item{fixed}{Use fixed (non-regex) matching instead of regular expressions.}
}
\value{
A list of character vectors.
}
\description{
Split text on a separator
}
