\name{transliterate}
\alias{transliterate}
\title{Transliterate text}
\usage{
transliterate(x, rule = "Any-Latin; Latin-ASCII")
}
\arguments{
\item{x}{A character vector or factor.}

\item{rule}{A stringi transliteration rule.}
}
\value{
A character vector transliterated according to the requested rule.
}
\description{
Transliterate text using stringi.
}
