% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primitives.R
\name{transliterate}
\alias{transliterate}
\title{Transliterate text to ASCII}
\usage{
transliterate(x)
}
\arguments{
\item{x}{An atomic vector.}
}
\value{
A character vector.
}
\description{
Transliterate text to ASCII
}
