% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primitives.R
\name{unionrows}
\alias{unionrows}
\title{Set union of rows}
\usage{
unionrows(x, y, by = NULL)
}
\arguments{
\item{x}{An atomic vector.}

\item{y}{An atomic vector, data.frame, or data.table, depending on the function.}

\item{by}{Character vector of column names identifying groups or join keys.}
}
\value{
The union of \code{x} and \code{y}.
}
\description{
Set union of rows
}
