% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primitives.R
\name{uniquerows}
\alias{uniquerows}
\title{Unique rows}
\usage{
uniquerows(data, cols = NULL, .keep_all = FALSE)
}
\arguments{
\item{data}{A data.frame.}

\item{cols}{Optional character vector of columns used to determine
uniqueness.}

\item{.keep_all}{Keep all columns when \code{cols} is supplied.}
}
\value{
A basetable of the unique rows.
}
\description{
Return unique rows, optionally considering only selected columns.
}
